+++ src/store/index.js
@@ -2,12 +2,14 @@ import { configureStore } from '@reduxjs/toolkit'
 import cartReducer from './cartSlice'
 import authReducer from './authSlice'
 import wishlistReducer from './wishlistSlice'
+import productReducer from './productSlice'
 
 const store = configureStore({
   reducer: {
     cart: cartReducer,
     auth: authReducer,
     wishlist: wishlistReducer,
+    products: productReducer,
   },
 })